+++ components/generator/jvm/imagevector/src/test/kotlin/io/github/composegears/valkyrie/generator/jvm/imagevector/ImageVectorWithPreviewTest.kt
@@ -29,35 +29,12 @@ class ImageVectorWithPreviewTest(
             config = createConfig(
                 outputFormat = outputFormat,
                 generatePreview = true,
-                previewAnnotationType = PreviewAnnotationType.AndroidX,
             ),
         ).content
 
         val expected = outputFormat.toResourceText(
             pathToBackingProperty = "imagevector/kt/backing/WithoutPath.preview.androidx.kt",
             pathToLazyProperty = "imagevector/kt/lazy/WithoutPath.preview.androidx.kt",
-        )
-        assertThat(parserOutput.iconType).isEqualTo(XML)
-        assertThat(output).isEqualTo(expected)
-    }
-
-    @Test
-    fun `jetbrains preview generation without icon pack`() {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
-        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
-        val output = ImageVectorGenerator.convert(
-            vector = parserOutput.irImageVector,
-            iconName = parserOutput.iconName,
-            config = createConfig(
-                outputFormat = outputFormat,
-                generatePreview = true,
-                previewAnnotationType = PreviewAnnotationType.Jetbrains,
-            ),
-        ).content
-
-        val expected = outputFormat.toResourceText(
-            pathToBackingProperty = "imagevector/kt/backing/WithoutPath.preview.jetbrains.kt",
-            pathToLazyProperty = "imagevector/kt/lazy/WithoutPath.preview.jetbrains.kt",
         )
         assertThat(parserOutput.iconType).isEqualTo(XML)
         assertThat(output).isEqualTo(expected)
@@ -74,36 +51,12 @@ class ImageVectorWithPreviewTest(
                 packName = "ValkyrieIcons",
                 outputFormat = outputFormat,
                 generatePreview = true,
-                previewAnnotationType = PreviewAnnotationType.AndroidX,
             ),
         ).content
 
         val expected = outputFormat.toResourceText(
             pathToBackingProperty = "imagevector/kt/backing/WithoutPath.pack.preview.androidx.kt",
             pathToLazyProperty = "imagevector/kt/lazy/WithoutPath.pack.preview.androidx.kt",
-        )
-        assertThat(parserOutput.iconType).isEqualTo(XML)
-        assertThat(output).isEqualTo(expected)
-    }
-
-    @Test
-    fun `jetbrains preview generation with icon pack`() {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
-        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
-        val output = ImageVectorGenerator.convert(
-            vector = parserOutput.irImageVector,
-            iconName = parserOutput.iconName,
-            config = createConfig(
-                packName = "ValkyrieIcons",
-                outputFormat = outputFormat,
-                generatePreview = true,
-                previewAnnotationType = PreviewAnnotationType.Jetbrains,
-            ),
-        ).content
-
-        val expected = outputFormat.toResourceText(
-            pathToBackingProperty = "imagevector/kt/backing/WithoutPath.pack.preview.jetbrains.kt",
-            pathToLazyProperty = "imagevector/kt/lazy/WithoutPath.pack.preview.jetbrains.kt",
         )
         assertThat(parserOutput.iconType).isEqualTo(XML)
         assertThat(output).isEqualTo(expected)
@@ -121,37 +74,12 @@ class ImageVectorWithPreviewTest(
                 nestedPackName = "Filled",
                 outputFormat = outputFormat,
                 generatePreview = true,
-                previewAnnotationType = PreviewAnnotationType.AndroidX,
             ),
         ).content
 
         val expected = outputFormat.toResourceText(
             pathToBackingProperty = "imagevector/kt/backing/WithoutPath.pack.nested.preview.androidx.kt",
             pathToLazyProperty = "imagevector/kt/lazy/WithoutPath.pack.nested.preview.androidx.kt",
-        )
-        assertThat(parserOutput.iconType).isEqualTo(XML)
-        assertThat(output).isEqualTo(expected)
-    }
-
-    @Test
-    fun `jetbrains preview generation with nested pack`() {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
-        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
-        val output = ImageVectorGenerator.convert(
-            vector = parserOutput.irImageVector,
-            iconName = parserOutput.iconName,
-            config = createConfig(
-                packName = "ValkyrieIcons",
-                nestedPackName = "Filled",
-                outputFormat = outputFormat,
-                generatePreview = true,
-                previewAnnotationType = PreviewAnnotationType.Jetbrains,
-            ),
-        ).content
-
-        val expected = outputFormat.toResourceText(
-            pathToBackingProperty = "imagevector/kt/backing/WithoutPath.pack.nested.preview.jetbrains.kt",
-            pathToLazyProperty = "imagevector/kt/lazy/WithoutPath.pack.nested.preview.jetbrains.kt",
         )
         assertThat(parserOutput.iconType).isEqualTo(XML)
         assertThat(output).isEqualTo(expected)